Add basic pw command
When you run it, it runs apg with sensible arguments, then uses ncurses to show
the options. You can move the cursor up or down and choose which password you
like. Or hit q to cancel. Once you choose a password, it copies it to your
clipboard and writes it to ~/.pw/history.